Code Refactoring

+++ src/Application/CommandHandlers/DeleteEntityCommandHandler.cs
@@ -1,4 +1,6 @@
 ﻿using Application.Commands;
+using Common.Notifications;
+using Domain.Core.Implementation.Events;
 using Domain.Core.Interfaces.Infrastructure;
 using Domain.Entities.EntityAggregate;
 using MediatR;
@@ -9,17 +11,33 @@ namespace Application.CommandHandlers
 {
     public class DeleteEntityCommandHandler : IRequestHandler<DeleteEntityCommand, bool>
     {
+        private INotificationManager _notificationManager;
+        private IMediator _mediator;
         private IRepository<EntityDomain> _entityRepository;
 
-        public DeleteEntityCommandHandler(IRepository<EntityDomain> entityRepository)
+        public DeleteEntityCommandHandler(
+            INotificationManager notificationManager,
+            IMediator mediator,
+            IRepository<EntityDomain> entityRepository
+            )
         {
+            _notificationManager = notificationManager;
+            _mediator = mediator;
             _entityRepository = entityRepository;
         }
 
-        public Task<bool> Handle(DeleteEntityCommand request, CancellationToken cancellationToken)
+        public async Task<bool> Handle(DeleteEntityCommand request, CancellationToken cancellationToken)
         {
+            var entity = _entityRepository.GetById(request.Id);
+            if(entity == null)
+            {
+                _notificationManager.Errors.Add(new NotificationMessage("Entity not found."));
+                return false;
+            }
+
             _entityRepository.Delete(request.Id);
-            return Task.FromResult(true);
+            await _mediator.Publish(new EntityDeletedDomainEvent<EntityDomain>(entity));
+            return true;
         }
     }
 }
